Changed "vars" to explicit types & added a using to improve readability
In my view it is better to be explicit when possible instead of using "var" as readability is quite important. I read from left to right, and to see "var" instead of the explicit type (when it's 100% deterministic what type will come out of a method call) makes me have to read more per line for every file I need to read.

It might look cleaner, but it also loses some readability.