Introduced an ApiErrorCode enum
This is to make the code writing less error prone. Every api code is an integer in the codebase and can easily be forgotten, used multiple times for different error messages and so on. Instead of risking writing faulty error messages, the integers have been replaced by these enum values.

It could also be used to auto generate error text, given the enum value's name. Lastly, if an error code needs to change for any reason, you only have to change the integer that the enum points to and you are done. Easier to maintain and means reusing code is easier too.